+++ tests/integration/api/v2010/account/sip/test_domain.py
@@ -54,7 +54,8 @@ class DomainTestCase(IntegrationTestCase):
                         "voice_status_callback_method": "POST",
                         "voice_status_callback_url": null,
                         "voice_url": "https://dundermifflin.example.com/twilio/app.php",
-                        "emergency_calling_enabled": true
+                        "emergency_calling_enabled": true,
+                        "secure": true
                     }
                 ],
                 "start": 0,
@@ -140,7 +141,8 @@ class DomainTestCase(IntegrationTestCase):
                 "voice_status_callback_method": "POST",
                 "voice_status_callback_url": null,
                 "voice_url": "https://dundermifflin.example.com/twilio/app.php",
-                "emergency_calling_enabled": true
+                "emergency_calling_enabled": true,
+                "secure": true
             }
             '''
         ))
@@ -189,7 +191,8 @@ class DomainTestCase(IntegrationTestCase):
                 "voice_status_callback_method": "POST",
                 "voice_status_callback_url": null,
                 "voice_url": "https://dundermifflin.example.com/twilio/app.php",
-                "emergency_calling_enabled": true
+                "emergency_calling_enabled": true,
+                "secure": true
             }
             '''
         ))
@@ -238,7 +241,8 @@ class DomainTestCase(IntegrationTestCase):
                 "voice_status_callback_method": "POST",
                 "voice_status_callback_url": null,
                 "voice_url": "https://dundermifflin.example.com/twilio/app.php",
-                "emergency_calling_enabled": true
+                "emergency_calling_enabled": true,
+                "secure": true
             }
             '''
         ))