+++ twilio/rest/serverless/v1/service/environment/deployment.py
@@ -113,7 +113,7 @@ class DeploymentList(ListResource):
 
         return DeploymentPage(self._version, response, self._solution)
 
-    def create(self, build_sid):
+    def create(self, build_sid=values.unset):
         """
         Create the DeploymentInstance